testing

+++ Scripts/Event/ASIUA;PLANNING!~!~!~.js
@@ -12,7 +12,8 @@ try {
 		var countParcels = 0;
 		for (a in tempAsit) {
 			if (!isNaN(tempAsit[a]["Parcel Acreage"])) {
-				parcelAcreage += parseFloat(tempAsit[a]["Parcel Acreage"]);
+				//parcelAcreage += parseFloat(tempAsit[a]["Parcel Acreage"]);
+				parcelAcreage += parseInt(tempAsit[a]["Parcel Acreage"]);
 			}
 			if (!isNaN(tempAsit[a]["Revised Acreage"])) {
 				revisedAcreage += parseFloat(tempAsit[a]["Revised Acreage"]);
